requirements.txt
torch>=2.0.0
numpy>=1.22.0
Pillow>=9.0.0
onnxruntime>=1.14.0
ultralytics>=8.0.0
scipy>=1.8.0 